mm/mm_gran: fix data truncation by using size_t for mask
Use size_t instead of unsigned int for 'mask' to avoid address truncation,
especially when memory addresses exceed 32 bits.

+++ mm/mm_gran/mm_graninit.c
@@ -99,7 +99,7 @@ GRAN_HANDLE gran_initialize(FAR void *heapstart, size_t heapsize,
   FAR struct gran_s *priv;
   uintptr_t          heapend;
   uintptr_t          alignedstart;
-  unsigned int       mask;
+  size_t             mask;
   unsigned int       alignedsize;
   unsigned int       ngranules;
